requirements.txt
torch
torchvision
torchinfo
numpy
pandas
seaborn
pillow
tqdm
matplotlib
